// src/types/lapdata.ts
export interface LapData {
  id: string;
  fileFireStorage: string;
  name: string;
  track: string;
  laptime: number;
  car: string;
  timestamp: number;
  trackGrip: number;
  weather: string;
  airTemp: number;
  roadTemp: number;
  sessionType: number;
  rainTypes: number;
  lapNumber: string;
}